refactor: rename project name and add spotless plugin.
