+++ auth-hands-on/auth-hands-on-api/settings.gradle.kts
@@ -1,0 +1,1 @@
+rootProject.name = "authHandsonApi"
