+++ auth-hands-on/auth-hands-on-api/src/main/kotlin/com/minmin/authHandsonApi/AuthHansOnApiApplication.kt
@@ -1,0 +1,11 @@
+package com.minmin.authHandsonApi
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class AuthHansOnApiApplication
+
+fun main(args: Array<String>) {
+    runApplication<AuthHansOnApiApplication>(*args)
+}
