+++ auth-hands-on/auth-hands-on-api/src/main/resources/application.yaml
@@ -1,0 +1,5 @@
+spring:
+  application:
+    name: authHandsonApi
+server:
+  port: 8888
