+++ auth-hands-on/auth-hands-on-api/src/test/kotlin/com/minmin/auth_hans_on_api/AuthHansOnApiApplicationTests.kt
@@ -1,0 +1,11 @@
+package com.minmin.authHandsonApi
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class AuthHansOnApiApplicationTests {
+    @Test
+    fun contextLoads() {
+    }
+}
