+++ auth-hands-on/auth-hands-on-client/settings.gradle.kts
