+++ auth-hands-on/auth-hands-on-client/src/main/kotlin/com/minmin/authHandsOnClient/AuthHansOnClientApplication.kt
@@ -1,0 +1,11 @@
+package com.minmin.authHandsOnClient
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class AuthHansOnClientApplication
+
+fun main(args: Array<String>) {
+    runApplication<AuthHansOnClientApplication>(*args)
+}
