+++ auth-hands-on/auth-hands-on-client/src/main/kotlin/com/minmin/authHandsOnClient/dtos/AccessToken.kt
@@ -1,0 +1,7 @@
+package com.minmin.authHandsOnClient.dtos
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies
+import com.fasterxml.jackson.databind.annotation.JsonNaming
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
+class AccessToken : JsonWebToken()
