+++ auth-hands-on/auth-hands-on-client/src/main/kotlin/com/minmin/authHandsOnClient/dtos/IdToken.kt
@@ -1,0 +1,7 @@
+package com.minmin.authHandsOnClient.dtos
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies
+import com.fasterxml.jackson.databind.annotation.JsonNaming
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
+class IdToken(val scope: String) : JsonWebToken()
