+++ auth-hands-on/auth-hands-on-client/src/main/kotlin/com/minmin/authHandsOnClient/utils/Base64UrlUtil.kt
@@ -1,0 +1,24 @@
+package com.minmin.authHandsOnClient.utils
+
+import java.util.Base64
+
+class Base64UrlUtil {
+    companion object {
+        fun encode(octets: ByteArray): String {
+            var encoded = Base64.getEncoder().encodeToString(octets)
+            encoded = encoded.split("=")[0]
+            encoded = encoded.replace('+', '-')
+            encoded = encoded.replace('/', '_')
+            return encoded
+        }
+
+        fun decode(encodedContent: String): ByteArray {
+            var encodedContentVar = encodedContent.replace('-', '+').replace('_', '/')
+            when (encodedContentVar.length % 4) {
+                2 -> encodedContentVar += "=="
+                3 -> encodedContentVar += "="
+            }
+            return Base64.getDecoder().decode(encodedContentVar)
+        }
+    }
+}
