+++ auth-hands-on/auth-hands-on-client/src/main/resources/templates/index.html
