+++ auth-hands-on/auth-hands-on-client/src/test/kotlin/com/minmin/auth_hans_on_client/AuthHansOnClientApplicationTests.kt
@@ -1,0 +1,8 @@
+package com.minmin.authHandsOnClient
+
+// @SpringBootTest
+class AuthHansOnClientApplicationTests {
+    // @Test
+    // fun contextLoads() {
+    // }
+}
