--- auth-hans-on/auth-hans-on-api/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "auth-hans-on-api"
