--- auth-hans-on/auth-hans-on-api/src/main/kotlin/com/minmin/auth_hans_on_api/AuthHansOnApiApplication.kt
@@ -1,11 +1,0 @@
-package com.minmin.auth_hans_on_api
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class AuthHansOnApiApplication
-
-fun main(args: Array<String>) {
-	runApplication<AuthHansOnApiApplication>(*args)
-}
