--- auth-hans-on/auth-hans-on-api/src/main/kotlin/com/minmin/auth_hans_on_api/controllers/TokenIntrospectionController.kt
@@ -1,16 +1,0 @@
-package com.minmin.auth_hans_on_api.controllers
-
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("token-introspection")
-class TokenIntrospectionController {
-    @GetMapping("/health")
-    fun getHealth(): ResponseEntity<String> {
-
-        return ResponseEntity.ok("TokenIntrospection Controller is up and runinning!!")
-    }
-}
