--- auth-hans-on/auth-hans-on-api/src/main/resources/application.yaml
@@ -1,5 +1,0 @@
-spring:
-  application:
-    name: auth-hans-on-api
-server:
-  port: 8888
