--- auth-hans-on/auth-hans-on-api/src/test/kotlin/com/minmin/auth_hans_on_api/AuthHansOnApiApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.minmin.auth_hans_on_api
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class AuthHansOnApiApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
