--- auth-hans-on/auth-hans-on-client/src/main/kotlin/com/minmin/auth_hans_on_client/AuthHansOnClientApplication.kt
@@ -1,11 +1,0 @@
-package com.minmin.auth_hans_on_client
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class AuthHansOnClientApplication
-
-fun main(args: Array<String>) {
-	runApplication<AuthHansOnClientApplication>(*args)
-}
