--- auth-hans-on/auth-hans-on-client/src/main/kotlin/com/minmin/auth_hans_on_client/dtos/IdToken.kt
@@ -1,7 +1,0 @@
-package com.minmin.auth_hans_on_client.dtos
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-class IdToken(val scope: String) : JsonWebToken() {}
