--- auth-hans-on/auth-hans-on-client/src/main/kotlin/com/minmin/auth_hans_on_client/dtos/RefreshToken.kt
@@ -1,8 +1,0 @@
-package com.minmin.auth_hans_on_client.dtos
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-import com.minmin.auth_hans_on_client.dtos.JsonWebToken
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-class RefreshToken : JsonWebToken()
