--- auth-hans-on/auth-hans-on-client/src/test/kotlin/com/minmin/auth_hans_on_client/AuthHansOnClientApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.minmin.auth_hans_on_client
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-// @SpringBootTest
-class AuthHansOnClientApplicationTests {
-
-	// @Test
-	// fun contextLoads() {
-	// }
-
-}
